feat: cross-bank 65816 disassembly (PBR-relative) (#508)
* feat: cross-bank 65816 disassembly (PBR-relative)

The disassembly panel follows the program bank: a 65816 program running in a
bank != 0 disassembles its own bytes instead of bank 0 at the same offset.

- cpu: DisasmCPUAt/WalkBackAt read instruction bytes through a supplied bus
  (the decode is bank-independent; only operand reads carry the bank).
- dap: bankView (a 16-bit Bus over peekByte24 of a fixed bank) + a 24-bit
  disassemble reference; symbols/source/data ranges stay a bank-0 concept.
- tui: Source.Disassemble takes a uint32 anchor; syncDisasm anchors at PBR:PC
  for the 65816; the panel shows $BB:XXXX + a (bank $NN) title.

Refs #507

* docs: cross-bank disassembly — ADR 0013 D2, context

+++ internal/tui/source.go
@@ -126,7 +126,7 @@ type Source interface {
 	// disassembly panel renders from (issue #452). LocalSource disassembles
 	// the live core via its inproc server; RemoteSource disassembles the
 	// DAP-fed mirror via its own inproc server (no per-tick wire round-trip).
-	Disassemble(anchor uint16, above, below int) (DisasmSnapshot, error)
+	Disassemble(anchor uint32, above, below int) (DisasmSnapshot, error)
 
 	// SetDataBreakpoints (re)installs the memory watchpoint set on the source
 	// via DAP `setDataBreakpoints` (issue #471) so the server enforces them
@@ -312,7 +312,7 @@ func (s *LocalSource) SetBanked(b *cpu.Banked24) {
 
 // Disassemble renders instructions around anchor through the in-process DAP
 // server attached to the live core (issue #452).
-func (s *LocalSource) Disassemble(anchor uint16, above, below int) (DisasmSnapshot, error) {
+func (s *LocalSource) Disassemble(anchor uint32, above, below int) (DisasmSnapshot, error) {
 	if s.dapClient == nil {
 		return DisasmSnapshot{}, fmt.Errorf("local source: no dap client")
 	}
